+++ cellpy/utils/helpers.py
@@ -200,9 +200,9 @@ def update_journal_cellpy_data_dir(
 
 def make_new_cell():
     """create an empty CellpyCell object."""
-    warnings.warn(
-        "make_new_cell is deprecated, use CellpyCell.vacant instead", DeprecationWarning
-    )
+    from cellpy._deprecation import warn_once
+
+    warn_once("make_new_cell", "CellpyCell.vacant")
     new_cell = cellpy.cellreader.CellpyCell(initialize=True)
     return new_cell